requirements.txt
aiogram==3.22.0
openai
python-dotenv
